@@ -17,15 +17,14 @@
         }
     ],
     "require": {
-        "php": "^7.2",
-        "illuminate/console": "^6.0"
+        "php": "^8.0",
+        "illuminate/console": "^8.0"
     },
     "require-dev": {
-        "phpstan/phpstan": "^0.11",
         "graham-campbell/testbench": "^5.1",
         "graham-campbell/testbench-core": "^3.0.3",
         "mockery/mockery": "^1.0",
-        "sven/laravel-testing-utils": "^1.3",
+        "sven/laravel-view-assertions": "^1.0",
         "friendsofphp/php-cs-fixer": "^2.15",
         "adamwojs/php-cs-fixer-phpdoc-force-fqcn": "^1.0"
     },
@@ -40,8 +39,7 @@
         }
     },
     "scripts": {
-        "cs": "php-cs-fixer fix --verbose",
-        "stan": "phpstan analyse src/ tests/ --level=7"
+        "cs": "php-cs-fixer fix --verbose"
     },
     "extra": {
         "laravel": {
