@@ -7,7 +7,7 @@ use Sven\ArtisanView\ServiceProvider;
 
 abstract class TestCase extends AbstractPackageTestCase
 {
-    protected function getServiceProviderClass($app)
+    protected function getServiceProviderClass()
     {
         return ServiceProvider::class;
     }
